% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ij2ijv_i}
\alias{ij2ijv_i}
\title{Transform Repeated Matrix Indexes}
\usage{
ij2ijv_i(ir, jc)
}
\arguments{
\item{ir}{An integer vector, row indexes}

\item{jc}{An integer vector, column indexes}
}
\value{
A list with fields i, j and iv
}
\description{
Transforms a couple of index vectors ir and jc (ij of a sparse matrix)
with possibly repeated values into sparse indexes i,j and a vector of 1d indexes of non zero values.
The response can be then used for repeated creation of sparse
matrices with the same pattern by calling \code{iv2v()}
ir and jc are supposed to be sorted in increasing order, column-wise (ic runs first)
}
