% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{iv2v}
\alias{iv2v}
\title{Sum non Zero Repeated Values}
\usage{
iv2v(iv, v)
}
\arguments{
\item{iv}{An integer vector, obtained with \code{ij2ijv_i(...)$iv}}

\item{v}{A numeric vector}
}
\value{
Numeric vector
}
\description{
sum values in v according to possibly repeated indexes in iv
}
