% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jrhs_ff}
\alias{jrhs_ff}
\title{Update Matrix by a Cascade of Dot Product}
\usage{
jrhs_ff(jrhs, ff, xpfw)
}
\arguments{
\item{jrhs}{A sparse matrix of type slam}

\item{ff}{A sparse matrix of type slam}

\item{xpfw}{A numeric matrix}
}
\description{
Update Matrix by a Cascade of Dot Product
}
