% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{match_ij}
\alias{match_ij}
\title{Fast Match for Matrix Indexes}
\usage{
match_ij(ix, jx, ti, tj)
}
\arguments{
\item{ix}{An integer vector}

\item{jx}{An integer vector}

\item{ti}{An integer vector}

\item{tj}{An integer vector}
}
\value{
An integer vector
}
\description{
Match ix,jx-couple in ti,tj-table and return their 1-based positions (0 for non matched couples)
}
\examples{
match_ij(1:2, 1:2, 0:4, 0:4)
# [1] 2 3

}
