% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mult_bxxc}
\alias{mult_bxxc}
\title{Calculate Inplace a Series of Dot Product}
\usage{
mult_bxxc(a, b, c)
}
\arguments{
\item{a}{A dense array, the size of a is (nr_b, nc_c, ntico)}

\item{b}{A sparse matrix (cf. \link[slam]{simple_triplet_matrix}) of size (nr_b*ntico, nc_b) given by its fields
v, i, and j describing triplet storage.}

\item{c}{A dense array, the size of c is (ldc, nc_c, ntico), ldc must be >= ncol(b)}
}
\value{
None
}
\description{
Calculate Inplace a Series of Dot Product
}
