\name{multbxxc-package}
\alias{multbxxc-package}
\alias{multbxxc}
\alias{bop}
\alias{ij2ijv_i}
\alias{iv2v}
\alias{jrhs_ff}
\alias{match_ij}
\alias{mm_xpf}
\alias{mult_bxxc}
\alias{redim}
\alias{resize}
\alias{solve_ieu}
\docType{package}
\title{
  Auxiliary routines for influx_si software
}
\description{
  This package contains auxiliary routines for influx_si software. It is not intended to be used directly.
}
\author{
Serguei Sokol

Maintainer: Serguei Sokol <sokol@insa-toulouse.fr>
}
\references{
  Sokol et al. (2012) <doi:10.1093/bioinformatics/btr716>
}
\keyword{ influx_si }
