% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multbxxc.R
\docType{package}
\name{multbxxc}
\alias{multbxxc}
\alias{multbxxc-package}
\title{multbxxc: Auxiliary Routines for Influx Software}
\description{
The multbxxc package provides a series C++ function most often operating inplace
}
\section{keyword}{

    metabolic flux analysis (MFA)
}

\references{
Sokol et al. (2012) <doi:10.1093/bioinformatics/btr716>
}
\author{
Serguei Sokol
}
