% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{redim}
\alias{redim}
\title{New Dimensions}
\usage{
redim(x, di)
}
\arguments{
\item{x}{A numeric array}

\item{di}{An integer vector, new dimensions}
}
\description{
Write new dimension vector while keeping the old memory
}
