% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resize}
\alias{resize}
\title{New Dimensions with Resizing}
\usage{
resize(x_, di)
}
\arguments{
\item{x_}{A numeric array}

\item{di}{An integer vector, new dimensions}
}
\description{
Write new dimension vector while keeping the old memory if possible
New memory cannot be greater than the very first allocation
}
