\name{multbxxc-package}
\alias{multbxxc-package}
\alias{multbxxc}
\docType{package}
\title{
  Auxiliary routines for influx_si software
}
\description{
  This package contains auxiliary routines for influx_si software. It is not intended to be used directly.
}
\author{
Serguei Sokol

Maintainer: Serguei Sokol <sokol@insa-toulouse.fr>
}
\references{
  Sokol et al. (2012) <doi:10.1093/bioinformatics/btr716>
}
\keyword{ influx_si }
