% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{redim}
\alias{redim}
\title{New Dimensions}
\usage{
redim(x, di)
}
\arguments{
\item{x}{A numeric array}

\item{di}{An integer vector, new dimensions}
}
\value{
None
}
\description{
Write new dimension vector while keeping the old memory
}
\examples{
a=matrix(as.double(1:12), 6, 2)
redim(a, c(3, 4))
dim(a)
# [1] 3 4
}
