% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resize}
\alias{resize}
\title{New Dimensions with Resizing}
\usage{
resize(x_, di)
}
\arguments{
\item{x_}{A numeric array}

\item{di}{An integer vector, new dimensions}
}
\value{
None
}
\description{
Write new dimension vector while keeping the old memory if possible
New memory cannot be greater than the very first allocation
}
\examples{
a=matrix(as.double(1:12), 6, 2)
resize(a, c(2, 2))
a
#      [,1] [,2]
# [1,]    1    3
# [2,]    2    4
}
